pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"reference_harvester.providers.uspto" = ["resources/*.yaml"]

[project]
name = "reference-harvester"
version = "0.1.0"
description = "Spec-driven, multi-provider reference/metadata harvesting and export"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Todo" }]
license = { text = "MIT" }
dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "httpx>=0.27.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
gui = [
    "nicegui>=2.5.0",
    "streamlit>=1.31.0",
]

[project.scripts]
reference-harvester = "reference_harvester.cli.app:run"

[tool.ruff]
line-length = 79

[tool.ruff.format]
